cleaning non used files and renaming files
